refactor: datadog and event helpers
